Revert "Drop contract pass; non-fast versions did not do a great job reducing unnecessary vmovapd instructions, and a (now fixed) bug causd issue 25. Fixes 25."
This reverts commit 589f36edf2d75c38206e45ac50e5457b487f3a00.
While the contract pass might not really decrease the amount of vmovapd
instrctusions, it is also helpful to more accurately estimate the cost
of sequentional mul and add instructions. Until this is actually
implemented for LoopSet objects themselves, using the contract_pass
function on the Expr before converting it into a LoopSet is a convenient
hack.